requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
requests==2.32.3
python-dotenv==1.0.1
feedparser==6.0.11
APScheduler==3.10.4
authlib==1.6.3
bleach==6.1.0
gunicorn==22.0.0
psycopg2-binary==2.9.10
tweepy==4.16.0
pyjwt==2.8.0
parcel==2.7.0